[feat] 스터디 그룹 멤버 추가 api연결

+++ data/src/main/java/com/umc/data/dataSource/OrganizationDataSource.kt
@@ -67,5 +67,5 @@ interface OrganizationDataSource {
     suspend fun changeActiveGisu(gisuId: Int): ApiState<Unit>
 
     // PUT
-    suspend fun changeGroupMember(groupId: Int, request: ChallengerListRequest): ApiState<Unit>
+    suspend fun changeGroupMember(groupId: Long, request: ChallengerListRequest): ApiState<Unit>
 }
